% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.for.plot.R
\name{content.for.plot}
\alias{content.for.plot}
\title{Transcript Categorization With Sampling}
\usage{
content.for.plot(data)
}
\arguments{
\item{data}{The object created by one of the sep.pane() functions}
}
\description{
This function takes the list created by one of the sep.panel() functions, and reformats it into meaningful codes, as well as reshapes the data into a melted
dataframe that will be supplied to ggplot. This formatting only works for making content bar charts specific to those generated by the transcriptBarChart
function.
}
\examples{
content.for.plot(content.summary)
}
\keyword{melt}
