% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contentQC.R
\name{contentQC}
\alias{contentQC}
\title{Summarize transcript content into a tabular format and perform QC checks}
\usage{
contentQC(subtype, dataList, refDat, i)
}
\arguments{
\item{subtype}{vector of present options as they are coded in the .results obects}

\item{dataList}{List .results datasets (i.e. known.results, ntkg.results, ntng.results)}

\item{refDat}{rbinded data of the 3 datasets listed above with the transcripts labeled "Not present" removed. This should have 114982 transcripts.}

\item{i}{index indicating whether or not to summarize ALO data or ALL data, i = 1 for ALO, and 2 for ALL}
}
\description{
This function summarizes the transcript content and checks to make sure
each subcategory of transcripts is a true subset of the complete ALO set of transcripts.
}
\examples{
contentQC(subtype, dataList, refDat, i)
}
