% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon.tpt.table.R
\name{exon.tpt.table}
\alias{exon.tpt.table}
\title{Calculate Exons Per Transcript}
\usage{
exon.tpt.table(ST_df.known = NULL, ST_df.ntng, ST_df.ntkg, directory, chrSize)
}
\arguments{
\item{ST_df.known}{R version of gtf subsetted for known transcripts}

\item{ST_df.ntng}{R version of gtf subsetted for novel transcripts of novel genes}

\item{ST_df.ntkg}{R version of gtf subsetted for novel transcripts of known genes}

\item{directory}{directory pointing to chromosome sizes file}

\item{chrSizes}{name of chromosome sizes file}

\item{tissue}{name of tissue (this has to do with the way my files are structured)}
}
\description{
Brief Description: This function takes as input the cleaned
chromosome file and the data frames created from the merged
liver gtf, which have been subsetted based on transcripts 
present in the known, ntng or ntkg data. It then creates a reference
table of transcript names and exons, and calculates the 
number of exons/transcript.
}
\keyword{chromosome}
\keyword{gtf}
\keyword{sizes}
