% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{ImportSeurat}
\alias{ImportSeurat}
\title{Import Seurat object to StcrSet}
\usage{
ImportSeurat(sset, seuratObject, version = "v2")
}
\arguments{
\item{sset}{StcrSet object which contains cellName}

\item{seuratObject}{Second StcrSet which contains cellName}

\item{version}{Seurat object version (default v2). v3 is also available.}
}
\description{
Import Seurat object to StcrSet
}
\examples{
\dontrun{
sset_NKT = ImportSeurat(sset,seurat.NKT)
}
}
