% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{Load10xVDJ}
\alias{Load10xVDJ}
\title{Load contig annotation file from 10X(>2.2.0)}
\usage{
Load10xVDJ(dataDir = NULL, filtered = TRUE, highConf = TRUE,
  productive = TRUE)
}
\arguments{
\item{dataDir}{Directory containing contig_annotations.csv}

\item{filtered}{boolean option to choose filtered(TRUE)/all(FALSE)}

\item{highConf}{boolean option to set highConfidence}

\item{productive}{boolean option to set productive}
}
\description{
Opens vdj result file provided by 10X genomics.
returns data.frame
}
\examples{
\dontrun{
data_dir = 'path/to/data/directory'
vdj_contig = StcrLoad(dataDir = data_dir, filtered = FALSE)
}
}
