% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{MergeSamples}
\alias{MergeSamples}
\title{merge two ssets}
\usage{
MergeSamples(sset1, sset2)
}
\arguments{
\item{sset1}{First StcrSet which contains cellName}

\item{sset2}{Second StcrSet which contains cellName}
}
\description{
merge two ssets
}
\examples{
\dontrun{
sset = MergeSamples(ssetF3,ssetF5) 
}
}
