% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\name{PlotClonotypes}
\alias{PlotClonotypes}
\title{Plot of frequencies by cluster}
\usage{
PlotClonotypes(CDR3PairMat, ident, cellLimit = 1)
}
\arguments{
\item{ident}{It can be cluster or celltype}

\item{cellLimit}{minimum paired cell number}
}
\description{
Plot of frequencies by cluster
}
\examples{
\dontrun{
sset_NKT = PlotTCR(sset, group.by = )
}
}
