% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\name{PlotDiversity}
\alias{PlotDiversity}
\title{Plot of frequencies by cluster}
\usage{
PlotDiversity(CDR3PairMat, ident = NULL)
}
\arguments{
\item{ident}{It can be cluster or celltype}
}
\description{
Plot of frequencies by cluster
}
\examples{
\dontrun{
sset_NKT = PlotTCR(sset, group.by = )
}
}
