% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\name{PlotFancyChainUsage}
\alias{PlotFancyChainUsage}
\title{Plot of frequencies by cluster}
\usage{
PlotFancyChainUsage(CDR3Mat, cellLimit = 3)
}
\arguments{
\item{cellLimit}{minimum paired cell number}

\item{ident}{It can be cluster or celltype}
}
\description{
plot was inspired by vdjtools(https://github.com/mikessh/vdjtools)
}
\examples{
\dontrun{
sset_NKT = PlotTCR(sset, group.by = )
}
}
