% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\name{PlotFrequency}
\alias{PlotFrequency}
\title{Plot of (non)canonical frequencies by cluster}
\usage{
PlotFrequency(CDR3PairMat, ident = NULL, reverse = TRUE,
  selected_ind = NULL, chain1_v_gene = NULL, chain1_d_gene = NULL,
  chain1_j_gene = NULL, chain2_v_gene = NULL, chain2_d_gene = NULL,
  chain2_j_gene = NULL)
}
\arguments{
\item{CDR3PairMat}{}

\item{ident}{It can be cluster or celltype}

\item{chain1_v_gene}{ex) c("TRAV11","TRAJ18"...) used to assign canonical and noncanonical}

\item{chain1_d_gene}{regex string for chain1_d_gene. default value is NULL. ex "TRAJ18"}

\item{chain1_j_gene}{regex string for chain1_j_gene. default value is NULL. ex "TRAV11"}

\item{chain2_v_gene}{regex string for chain2_v_gene. default value is NULL. ex "(TRBV13|TRBV29|TRBV1\\*)"}

\item{chain2_d_gene}{regex string for chain2_d_gene. default value is NULL.}

\item{chain2_j_gene}{regex string for chain2_j_gene. default value is NULL.}

\item{canonical}{frequency of canonical TCRs if the value is TRUE. default value is TRUE}

\item{canonical_ind}{a vector of canonical cell index}
}
\description{
PlotFrequency takes either canonical_ind or gene regex string from chains and
returns a ggplot object of (non)canonical frequencies in each of the clusters.
All gene regex string are combinded with "AND" operation to assign canonical_ind.
}
\examples{
\dontrun{

test_cdr3_pair_mat = SetCDR3PairMat(sset_test, test_cdr3_mat, chain1 = c("TRA"), chain2 = c("TRB"))
PlotFrequency <- function(test_cdr3_pair_mat, ident = sset_test@ident , canonical_ind= NULL,
 chain1_v_gene = NULL, chain1_d_gene = NULL, chain1_j_gene = NULL, 
 chain2_v_gene = NULL, chain2_d_gene = NULL, chain2_j_gene = NULL)
If you have a list of canonical cell index
PlotFrequency <- function(test_cdr3_pair_mat, ident = sset_test@ident , canonical_ind= canonical_ind)

}
}
