% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\name{PlotPairUsageHeatmap}
\alias{PlotPairUsageHeatmap}
\title{Plot of PairUsage}
\usage{
PlotPairUsageHeatmap(mat_1 = NULL, mat_2 = NULL, cell_ind = NULL,
  pseudocount = 0.1)
}
\arguments{
\item{pseudocount}{}
}
\description{
Plot of PairUsage
}
\examples{
\dontrun{
sset_NKT = PlotTCR(sset, group.by = )
}
}
