% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\name{PlotProportionsByGene}
\alias{PlotProportionsByGene}
\title{Bar plot of frequencies by cluster}
\usage{
PlotProportionsByGene(CDR3PairMat, ident = NULL, selected = TRUE,
  selected_ind = NULL, chain1_v_gene = NULL, chain1_d_gene = NULL,
  chain1_j_gene = NULL, chain2_v_gene = NULL, chain2_d_gene = NULL,
  chain2_j_gene = NULL)
}
\arguments{
\item{ident}{It can be cluster or celltype}

\item{chain1_v_gene}{ex) c("TRAV11","TRAJ18"...) used to assign canonical and noncanonical}

\item{chain1_d_gene}{regex string for chain1_d_gene. default value is NULL. ex "TRAJ18"}

\item{chain1_j_gene}{regex string for chain1_j_gene. default value is NULL. ex "TRAV11"}

\item{chain2_v_gene}{regex string for chain2_v_gene. default value is NULL. ex "(TRBV13|TRBV29|TRBV1\\*)"}

\item{chain2_d_gene}{regex string for chain2_d_gene. default value is NULL.}

\item{chain2_j_gene}{regex string for chain2_j_gene. default value is NULL.}

\item{canonical_cell_ind}{a vector of canonical cell index}
}
\description{
Canonical/(Non)canoncial Frequency plot
}
\examples{
\dontrun{
sset_NKT = PlotTCR(sset, group.by = )
}
}
