% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\name{PlotTCR}
\alias{PlotTCR}
\title{Scatter plot of a binary meta.data on a reduced dimension}
\usage{
PlotTCR(sset, group_by = NULL, cell_ind = NULL, reduction = "umap",
  color1 = "navy", color2 = "grey")
}
\arguments{
\item{sset}{StcrSet object which contains cellName}

\item{group_by}{uses meta.data of cell from sset}

\item{cell_ind}{A vector of cells to be highlighted}

\item{reduction}{Which dimensionality reduction to use.}

\item{color1}{primary color, default navy}

\item{color2}{backgound cell color, default grey}
}
\description{
Scatter plot of a binary meta.data on a reduced dimension
}
\examples{
\dontrun{
plt = PlotTCR(sset, group_by = )
print(plt)
plt = PlotTCR(sset, group_by = )
print(plt)
}
}
