% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pres.abs.table.R
\name{pres.abs.table}
\alias{pres.abs.table}
\title{Generate Present/Absent Table}
\usage{
pres.abs.table(cnts, datanames, presThresh = 10, absThresh = 1)
}
\arguments{
\item{cnts}{expected count matrix where rows are transcript/gene ids and columns are strains. The first column should be a vector of transcript ids}

\item{datanames}{Character vector of unique strain names used in cnts data including transcript id col name}

\item{presThresh}{Upper count threshold to be considered present (default = 10)}

\item{absThresh}{Lower count threshold to be considered absent (default = 1)}
}
\description{
This function identifies transcripts as present/absent, or other based on a user defined
criteria, and then assigns a value (1 =Present, 0=Absent, 0.5 =Other) on the 
strain level.
}
\examples{
pres.abs.table(cnts, names, 10, 1)
}
\keyword{Present/Absent}
