% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.read.iso.R
\name{rsem.read.iso}
\alias{rsem.read.iso}
\title{Reading in RSEM output - Isoform}
\usage{
rsem.read.iso(dir, dirs2, tissue, N, qtype, level, location, filename)
}
\arguments{
\item{dir}{Global directory. Often points to project start folder. This will be the Parent folder for the folder containing lists of file names for each rsem batch output}

\item{dirs2}{Directory pointing parent folder where RSEM output is}

\item{tissue}{Character value defining name of tissue in file path}

\item{N}{Numerical vector of batch numbers}

\item{qtype}{What quantitation type should the function read in. Can read counts or TPM}

\item{level}{Should the function generate expression data at the gene level or transcript level}

\item{location}{name of folder that holds target .results files}

\item{filename}{file name for output as RData file}
}
\description{
This function reads in the RSEM output for each sample and merges the data into one count matrix with rows as transcript IDs and columns as sample names
}
\examples{
rsem.read.iso()
}
\keyword{rsem}
