% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{SetCDR3Mat}
\alias{SetCDR3Mat}
\title{Create a matrix of CDR3 of each chains}
\usage{
SetCDR3Mat(sset, cell_ind = NULL, chain1 = c("TRA"),
  chain2 = c("TRB"), v_gene1 = NULL, v_gene2 = NULL,
  j_gene1 = NULL, j_gene2 = NULL)
}
\arguments{
\item{sset}{StcrSet object which contains cellName}

\item{chain1}{chains(more than one can also be possible)}

\item{chain2}{chains(more than one can also be possible)}

\item{v_gene1}{additional gene info, if you use more than one chain}

\item{v_gene2}{additional gene info, if you use more than one chain}

\item{j_gene1}{additional gene info, if you use more than one chain}

\item{j_gene2}{additional gene info, if you use more than one chain}
}
\description{
chain1 -> row, chian2 -> column
VJ
return a matrix
}
\examples{
\dontrun{
gdT_cdr3_mat = SetCDR3Mat(sset_gdT, cell_ind = colnames(seurat.gdT@scale.data), 
chain1 = c("TRG"), chain2 = c("TRD","Multi"), v_gene2 = "DV", j_gene2="DJ")
}
}
