% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{SetCDR3PairMat}
\alias{SetCDR3PairMat}
\title{Create a matrix of CDR3 of each chains}
\usage{
SetCDR3PairMat(sset, CDR3Mat, chain1 = c("TRA"), chain2 = c("TRB"))
}
\arguments{
\item{sset}{StcrSet object which contains cellName}

\item{CDR3Mat}{a list output of SetCDR3Mat}

\item{chain1}{default TRA}

\item{chain2}{default TRB}
}
\description{
chain1 -> row, chian2 -> column
return a list with a matrix of two CDR3 chains and a matrix of its CDR3 pair and clonotypes
}
\examples{
\dontrun{
gdT_cdr3_mat = SetCDR3Mat(sset_gdT, cell_ind = colnames(seurat.gdT@scale.data), 
chain1 = c("TRG"), chain2 = c("TRD","Multi"), v_gene2 = "DV", j_gene2="DJ")
gdT_cdr3_pair_mat = SetCDR3PairMat(sset_gdT, gdT_cdr3_mat, chain1 = c("TRG"), chain2 = c("TRD","Multi"))
}
}
