% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{SetSample}
\alias{SetSample}
\title{Assign sample ID to a sset}
\usage{
SetSample(sset, sampleID = NULL, sep = "-")
}
\arguments{
\item{sset}{StcrSet}

\item{sampleID}{string}
}
\description{
Assign sample ID to a sset
}
\examples{
\dontrun{
ssetF3 = SetSample(ssetF3,"F3")
}
}
