% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{SetVDJcountMat}
\alias{SetVDJcountMat}
\title{Create a matrix using StcrSet and one of vdj gene.}
\usage{
SetVDJcountMat(sset, chain = NULL, ref = "IMGT", v_gene = NULL,
  d_gene = NULL, j_gene = NULL)
}
\arguments{
\item{sset}{StcrSet object which contains cellName}

\item{chain}{a set of string contains more than one chain ex) c("TRA") or "TRA"}

\item{ref}{default "IMGT"}

\item{v_gene}{only one gene can be selected among v_gene, d_gene, j_gene ex)"AV"}

\item{d_gene}{only one gene can be selected among v_gene, d_gene, j_gene ex)"AD"}

\item{j_gene}{only one gene can be selected among v_gene, d_gene, j_gene ex)"AJ"}
}
\description{
return a matrix with specific chain and gene
}
\examples{
\dontrun{
mat_AV = SetVDJcountMat(sset_NKT, chain = "TRA", v_gene = "AV")
}
}
