\name{Stcr-package}
\alias{Stcr-package}
\alias{Stcr}
\docType{package}
\title{
  Single cell tcr analysis tool
}
\description{
  A more detailed description of what the package does. A length
  of about one to five lines is recommended.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Eunmin Lee, Eunseo Park, JongKyung Kim

Maintainer: Eunmin Lee <elee4472@dgist.ac.kr>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}

