% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSets.R
\docType{class}
\name{StcrSet}
\alias{StcrSet}
\alias{StcrSet-class}
\title{The stcrSet class}
\description{
The stcrSet object is the main class object that contains vdj results of 10x vdj.
We recommands to use IMGT reference and Seurat package to analyze the single cell RNA sequence data.
}
\section{Slots}{

\describe{
\item{\code{contig}}{a data frame of 10x contig}

\item{\code{cellName}}{a vector of cell names}

\item{\code{ident}}{a factor of cell types or clusters}

\item{\code{reductions}}{a list of reductions, it has the same format as Seurat object}

\item{\code{vdjMat}}{a list of}

\item{\code{meta.data}}{a data frame of meta.data from Seurat object}
}}

