% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.position.bytransc.R
\name{track.position.bytransc}
\alias{track.position.bytransc}
\title{Find Chromosome position for use in UCSC track visualizer with geneID}
\usage{
track.position.bytransc(eptOut, tlist1, tlist2, tlist3)
}
\arguments{
\item{eptOut}{output object from exon.tpt.table()}

\item{tlist1}{transcript list 1}

\item{tlist2}{transcript list 2}

\item{tlist3}{transcript list 3}
}
\description{
Brief Description: This function does the same thing as track.position.table() but allows the user to search for transcripts from a
gene ID chosen by the user a priori. These transcript lists go into the the function in place of the number of exons for the track.position.table function.
}
\examples{
track.position.bytransc(eptOut, tlist1, tlist2, tlist2)
}
\keyword{Genome}
\keyword{UCSC}
\keyword{browser}
\keyword{custome}
\keyword{track}
