% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptBarChart.R
\name{transcriptBarChart}
\alias{transcriptBarChart}
\title{Plot Transcript Content}
\usage{
transcriptBarChart(content.plot.data, directory, filename)
}
\arguments{
\item{data}{The object created by the content.for.plot()}
}
\description{
This function takes the output from the content.for.plot() function and plots the transcript content that was summarized. It generates 4 barcharts. 
A low and high resolution for the "At least one" criteria, and a low and high resolution for the "in all" criteria.
}
\examples{
transcriptBarChart(content.summary)
}
\keyword{barchart}
